requirements.txt
beautifulsoup4==4.12.3
langchain==0.2.14
langchain_community==0.2.12
langchain_core==0.2.32
langchain_openai==0.1.21
numpy==1.24.3
openai==1.40.8
prompt_toolkit==3.0.43
pydantic==1.10.9
python-dotenv==1.0.1
Requests==2.32.3
tiktoken==0.3.3
trafilatura==1.8.1
